store/wizard: move compliance selectors to submodule
