+++ src/store/slices/wizard/compliance/index.ts
@@ -1,3 +1,4 @@
+export * from './selectors';
 export { initialState as complianceState } from './state';
 export * from './types';
 export * from './utilities';
